update scripts for iFinder Core

+++ bootstrap/start-sitesearch-elasticsearch.ps1
@@ -11,15 +11,19 @@ $persistence_dir = "srv"
 
 docker network create $docker_network
 
-docker load -i "~/tmp/$service_name.tar"
+if(Test-Path ~/tmp/$service_name.tar){
+    docker load -i "~/tmp/$service_name.tar"
 
-mkdir -p $HOME/$persistence_dir/$service_name
-sudo chown -R 1000:1000 $HOME/$persistence_dir/$service_name
-sudo chmod -R 744 $HOME/$persistence_dir/$service_name
+    mkdir -p $HOME/$persistence_dir/$service_name
+    sudo chown -R 1000:1000 $HOME/$persistence_dir/$service_name
+    sudo chmod -R 744 $HOME/$persistence_dir/$service_name
 
-docker rm -f $service_name
-docker run -d --name $service_name `
-    --network $docker_network `
-    -v $HOME/$persistence_dir/$service_name/data:/home/app_user/data:rw `
-    -v $HOME/$persistence_dir/$service_name/logs:/home/app_user/logs:rw `
-    $docker_image
+    docker rm -f $service_name
+    docker run -d --name $service_name `
+        --network $docker_network `
+        -v $HOME/$persistence_dir/$service_name/data:/home/app_user/data:rw `
+        -v $HOME/$persistence_dir/$service_name/logs:/home/app_user/logs:rw `
+        $docker_image
+} else {
+    echo ~/tmp/$service_name.tar file is missing
+}
